// taurus-protect-sdk-typescript/src/internal/openapi/models/TgvalidatordETHValidatorInfo.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Taurus-PROTECT APIs
 * Taurus-PROTECT APIs description
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface TgvalidatordETHValidatorInfo
 */
export interface TgvalidatordETHValidatorInfo {
    /**
     * 
     * @type {string}
     * @memberof TgvalidatordETHValidatorInfo
     */
    id?: string;
    /**
     * 
     * @type {string}
     * @memberof TgvalidatordETHValidatorInfo
     */
    pubkey?: string;
    /**
     * 
     * @type {string}
     * @memberof TgvalidatordETHValidatorInfo
     */
    status?: string;
    /**
     * 
     * @type {string}
     * @memberof TgvalidatordETHValidatorInfo
     */
    balance?: string;
    /**
     * 
     * @type {string}
     * @memberof TgvalidatordETHValidatorInfo
     */
    network?: string;
    /**
     * 
     * @type {string}
     * @memberof TgvalidatordETHValidatorInfo
     */
    provider?: string;
    /**
     * 
     * @type {string}
     * @memberof TgvalidatordETHValidatorInfo
     */
    addressID?: string;
}

/**
 * Check if a given object implements the TgvalidatordETHValidatorInfo interface.
 */
export function instanceOfTgvalidatordETHValidatorInfo(value: object): value is TgvalidatordETHValidatorInfo {
    return true;
}

export function TgvalidatordETHValidatorInfoFromJSON(json: any): TgvalidatordETHValidatorInfo {
    return TgvalidatordETHValidatorInfoFromJSONTyped(json, false);
}

export function TgvalidatordETHValidatorInfoFromJSONTyped(json: any, ignoreDiscriminator: boolean): TgvalidatordETHValidatorInfo {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'pubkey': json['pubkey'] == null ? undefined : json['pubkey'],
        'status': json['status'] == null ? undefined : json['status'],
        'balance': json['balance'] == null ? undefined : json['balance'],
        'network': json['network'] == null ? undefined : json['network'],
        'provider': json['provider'] == null ? undefined : json['provider'],
        'addressID': json['addressID'] == null ? undefined : json['addressID'],
    };
}

  export function TgvalidatordETHValidatorInfoToJSON(json: any): TgvalidatordETHValidatorInfo {
      return TgvalidatordETHValidatorInfoToJSONTyped(json, false);
  }

  export function TgvalidatordETHValidatorInfoToJSONTyped(value?: TgvalidatordETHValidatorInfo | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'pubkey': value['pubkey'],
        'status': value['status'],
        'balance': value['balance'],
        'network': value['network'],
        'provider': value['provider'],
        'addressID': value['addressID'],
    };
}